% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fish_stats.R
\name{calc_fish_stats}
\alias{calc_fish_stats}
\title{calc_fish_stats}
\usage{
calc_fish_stats(count = fish_count, price = fish_price,
  graph = FALSE)
}
\arguments{
\item{count}{this is a table of fish count, default is fish_count}

\item{price}{this is a table of prices for fish, default is fish_price}

\item{graph}{does the user want a graph}
}
\value{
most common fish at each location

total revenue at each location

total revenue across all locations

a statement stating the total revenue, if graph is true

a graph of revenue by location
}
\description{
compute stuff about fish
}
